+++ Libplanet/Blockchain/NullChainStates.cs
@@ -37,7 +37,7 @@ namespace Libplanet.Blockchain
             GetAccount(offset).GetValidatorSet();
 
         public IAccount GetAccount(BlockHash? offset) =>
-            new Account(offset ?? default, new MerkleTrie(_keyValueStore));
+            new Account(new MerkleTrie(_keyValueStore));
 
         public IRecordableTrie CastToRecordableTrie(ITrie trie)
         {